+++ app/src/main/java/woowacourse/movie/feature/bookingseat/view/BookingSeatActivity.kt
@@ -217,6 +217,7 @@ class BookingSeatActivity :
         alarmManager.setExactAndAllowWhileIdle(
             AlarmManager.RTC_WAKEUP,
             alarmTime.toEpochSecond() * 1000,
+//            System.currentTimeMillis() * 1000,
             pendingIntent,
         )
     }
